// frontend/src/services/api/types.ts
export type BoardPromoImageItem = {
  id?: number
  name?: string
  url: string
}

export type BoardPromoVideoItem = {
  id?: number
  name?: string
  playEvery?: number
  url: string
}

export type BoardPromoMedia = {
  imageId?: number
  imageItems?: BoardPromoImageItem[]
  imageName?: string
  imageUrl?: string
  updatedAt?: string
  videoId?: number
  videoItems?: BoardPromoVideoItem[]
  videoName?: string
  videoUrl?: string
}